#5560 move modelData mappings to separate modules
